api/requirements.txt
flask
flask-restful
flask-cors
flask-cors
flask-sqlalchemy
Flask-PyMemcache
Flask-OpenID

marshmallow
click==7.1.2
gunicorn
celery
jmespath
toolz
python-dotenv
psycopg2-binary